+++ xmpp-client/src/main/java/org/xmpp/extension/bytestreams/s5b/S5bEvent.java
@@ -63,7 +63,7 @@ final class S5bEvent extends ByteStreamEvent {
             // 6.3.3 Target Acknowledges Bytestream
             IQ result = iq.createResult();
             result.setExtension(Socks5ByteStream.streamHostUsed(s5bSession.getStreamHost()));
-            xmppSession.send(iq);
+            xmppSession.send(result);
             return s5bSession;
         } catch (IOException e) {
             // If the Target tries but is unable to connect to any of the StreamHosts and it does not wish to attempt a connection from its side, it MUST return an <item-not-found/> error to the Requester.