+++ xmpp-core/src/main/java/org/xmpp/extension/compress/Failure.java
@@ -29,6 +29,7 @@ import org.xmpp.stream.ServerStreamElement;
 import javax.xml.bind.annotation.XmlElementRef;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlSeeAlso;
+import javax.xml.bind.annotation.XmlTransient;
 
 /**
  * The implementation of the {@code <failure/>} element in the {@code http://jabber.org/protocol/compress} namespace, which indicates failure during compression negotiation.
@@ -51,7 +52,11 @@ public final class Failure implements ServerStreamElement {
     }
 
     static abstract class Condition {
-        private final String name;
+
+        private String name;
+
+        private Condition() {
+        }
 
         protected Condition(String name) {
             this.name = name;